Changed to int

+++ src/main/java/org/bboxdb/tools/experiments/tuplestore/AbstractJDBCTupleStore.java
@@ -63,7 +63,7 @@ public abstract class AbstractJDBCTupleStore implements TupleStore {
 		final byte[] tupleBytes = TupleHelper.tupleToBytes(tuple);
 		final InputStream is = new ByteArrayInputStream(tupleBytes);
 		
-		insertStatement.setLong(1, Integer.parseInt(tuple.getKey() + 1));
+		insertStatement.setInt(1, Integer.parseInt(tuple.getKey() + 1));
 		insertStatement.setBlob(2, is);
 		insertStatement.execute();
 	
@@ -75,7 +75,7 @@ public abstract class AbstractJDBCTupleStore implements TupleStore {
 	@Override
 	public Tuple readTuple(final String key) throws Exception {
 
-		selectStatement.setLong(1, Integer.parseInt(key) + 1);
+		selectStatement.setInt(1, Integer.parseInt(key) + 1);
 		final ResultSet result = selectStatement.executeQuery();
 		
 		if(! result.next()) {
